+++ tests/test_core.py
@@ -1,4 +1,5 @@
 import os
+import logging
 import pytest
 import re
 import datetime
@@ -134,6 +135,21 @@ def test_init_datetime_subclass_long_timespan(test_pd_df):
 
     assert df["time"].max() == tmax
     assert df["time"].min() == tmin
+
+
+def test_init_empty_message(test_pd_df, caplog):
+    tdf = test_pd_df.copy()
+    tdf["extra_col"] = "test_val"
+    tdf["extra_col"] = np.nan
+
+    res = IamDataFrame(tdf)
+    assert "Primary Energy|Coal" not in res.variables().tolist()
+
+    drop_message = (
+        "Formatted data is empty! (perhaps there is a column full of nans?)"
+    )
+    message_idx = caplog.messages.index(drop_message)
+    assert caplog.records[message_idx].levelno == logging.WARNING
 
 
 def test_get_item(test_df):